+++ lib/td_dq/auth/pipeline/secure.ex
@@ -5,11 +5,6 @@ defmodule TdDq.Auth.Pipeline.Secure do
     error_handler: TdDq.Auth.ErrorHandler,
     module: TdDq.Auth.Guardian
 
-  # If there is a session token, validate it
-  # plug Guardian.Plug.VerifySession, claims: %{"typ" => "access"}
-  # If there is an authorization header, validate it
-  # plug Guardian.Plug.VerifyHeader, claims: %{"typ" => "access"}
-  # Load the user if either of the verifications worked
   plug Guardian.Plug.EnsureAuthenticated
   plug TdDq.Auth.Plug.CurrentResource
 end