+++ test/support/conn_case.ex
@@ -58,12 +58,14 @@ defmodule TdDqWeb.ConnCase do
 
     cond do
       tags[:admin_authenticated] ->
-        user = create_user(@admin_user_name, is_admin: true)
-        create_user_auth_conn(user)
+        @admin_user_name
+        |> create_claims(role: "admin")
+        |> create_user_auth_conn()
 
       user_name = tags[:authenticated_user] ->
-        user = create_user(user_name, is_admin: false)
-        create_user_auth_conn(user, tags[:role])
+        user_name
+        |> create_claims()
+        |> create_user_auth_conn(tags[:role])
 
       true ->
         {:ok, conn: ConnTest.build_conn()}