+++ lib/td_dq/rules/implementations.ex
@@ -283,12 +283,10 @@ defmodule TdDq.Rules.Implementations do
   end
 
   defp deleted_implementations(query, options, order) do
-    case Keyword.get(options, :deleted, false) do
-      true ->
-        with_deleted(query, order)
-
-      false ->
-        without_deleted(query, order)
+    if Keyword.get(options, :deleted, false) do
+      with_deleted(query, order)
+    else
+      without_deleted(query, order)
     end
   end
 
@@ -296,16 +294,8 @@ defmodule TdDq.Rules.Implementations do
     where(query, [ri, _r], not is_nil(ri.deleted_at))
   end
 
-  defp with_deleted(query, :results) do
-    where(query, [_r, ri], not is_nil(ri.deleted_at))
-  end
-
   defp without_deleted(query, :implementations) do
     where(query, [ri, _r], is_nil(ri.deleted_at))
-  end
-
-  defp without_deleted(query, :results) do
-    where(query, [_r, ri], is_nil(ri.deleted_at))
   end
 
   def enrich_system(
@@ -459,8 +449,14 @@ defmodule TdDq.Rules.Implementations do
   defp structure_ids([_ | _] = values), do: Enum.flat_map(values, &structure_ids/1)
   defp structure_ids(%Structure{id: id}), do: [id]
 
+  defp structure_ids(%ConditionRow{structure: structure, value: nil}),
+    do: structure_ids(structure)
+
   defp structure_ids(%ConditionRow{structure: structure, value: values}),
     do: structure_ids([structure | values])
+
+  defp structure_ids(%DatasetRow{structure: structure, clauses: nil}),
+    do: structure_ids(structure)
 
   defp structure_ids(%DatasetRow{structure: structure, clauses: clauses}),
     do: structure_ids([structure | clauses])