+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog
 
-## [Unreleased] 2021-01-11
+## [Unreleased]
 
 ### Added
 
 - [TD-3187] Filter executions by source and `PENDING` status
+
+### Changed
+
+- [TD-3163] Auth tokens now include `role` claim instead of `is_admin` flag
+- [TD-3182] Allow to use redis with password
 
 ## [4.11.0] 2021-01-11