+++ CHANGELOG.md
@@ -11,6 +11,15 @@
 - [TD-3163] Auth tokens now include `role` claim instead of `is_admin` flag
 - [TD-3182] Allow to use redis with password
 
+## [4.11.1] 2021-01-20
+
+### Fixed
+
+- [TD-3206] Added a little bit of defensive programming when parsing
+  implementation structure ids due to a migration that could make
+  `ConditionRow` elements not valid from the changeset point of view
+  while being functionaly workable.
+
 ## [4.11.0] 2021-01-11
 
 ### Added